+++ backend/baguni-common/src/main/java/baguni/common/event/events/LinkEvent.java
@@ -1,0 +1,13 @@
+package baguni.common.event.events;
+
+import lombok.Getter;
+
+@Getter
+public class LinkEvent extends Event {
+
+	private final String url;
+
+	public LinkEvent(String url) {
+		this.url = url;
+	}
+}
